# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AnchorPoint API",
    "version": "1.0.0",
    "description": "AnchorPoint Backend API for Stellar SEP-24 Anchor Operations",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "AnchorPoint Support",
      "url": "https://github.com/ceejaylaboratory/AnchorPoint"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authorization token obtained from /auth/token endpoint"
      }
    },
    "schemas": {
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Transaction unique identifier"
          },
          "userPublicKey": {
            "type": "string",
            "description": "User Stellar public key"
          },
          "assetCode": {
            "type": "string",
            "description": "Asset code (e.g., USD, EUR)"
          },
          "amount": {
            "type": "string",
            "description": "Transaction amount"
          },
          "kind": {
            "type": "string",
            "enum": [
              "deposit",
              "withdrawal"
            ],
            "description": "Transaction type"
          },
          "status": {
            "type": "string",
            "description": "Transaction status"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Transaction creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Transaction update timestamp"
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Total number of items"
          },
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "limit": {
            "type": "integer",
            "description": "Number of items per page"
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "MultisigTransaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "hash": {
            "type": "string"
          },
          "envelopeXdr": {
            "type": "string"
          },
          "creatorPublicKey": {
            "type": "string"
          },
          "requiredSigners": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "threshold": {
            "type": "integer"
          },
          "currentSignatures": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PARTIALLY_SIGNED",
              "READY",
              "SUBMITTED",
              "FAILED",
              "EXPIRED"
            ]
          },
          "signatures": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "signerPublicKey": {
                  "type": "string"
                },
                "signedAt": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "memo": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "stellarTxId": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/api/transactions": {
      "get": {
        "summary": "Get transaction history",
        "description": "Fetches transaction history for the authenticated user with pagination support.",
        "tags": [
          "Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "assetCode",
            "schema": {
              "type": "string"
            },
            "description": "Filter transactions by asset code (e.g., USDC, BTC)"
          },
          {
            "in": "query",
            "name": "sender",
            "schema": {
              "type": "string"
            },
            "description": "Search for transactions with matching sender metadata in indexed events"
          },
          {
            "in": "query",
            "name": "receiver",
            "schema": {
              "type": "string"
            },
            "description": "Search for transactions with matching receiver metadata in indexed events"
          },
          {
            "in": "query",
            "name": "memo",
            "schema": {
              "type": "string"
            },
            "description": "Search for transactions by memo or indexed event text"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transactions": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Transaction"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions/submit": {
      "post": {
        "summary": "Submit a pre-signed transaction",
        "description": "Validates and submits a pre-signed Stellar transaction XDR to the network.",
        "tags": [
          "Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "xdr"
                ],
                "properties": {
                  "xdr": {
                    "type": "string",
                    "description": "Base64 encoded transaction XDR"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "hash": {
                          "type": "string"
                        },
                        "ledger": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid transaction or disallowed operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/multisig/transactions": {
      "post": {
        "summary": "Create a new multisig transaction",
        "description": "Creates a new multisig transaction that requires signatures from multiple parties",
        "tags": [
          "Multisig"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "envelopeXdr",
                  "requiredSigners",
                  "threshold"
                ],
                "properties": {
                  "envelopeXdr": {
                    "type": "string",
                    "description": "Base64 encoded transaction envelope"
                  },
                  "requiredSigners": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Array of public keys that must sign"
                  },
                  "threshold": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Number of signatures required"
                  },
                  "memo": {
                    "type": "string",
                    "description": "Optional memo for the transaction"
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Optional expiration date"
                  },
                  "metadata": {
                    "type": "object",
                    "description": "Optional additional metadata"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transaction": {
                          "$ref": "#/components/schemas/MultisigTransaction"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "Get multisig transactions for the authenticated user",
        "description": "Retrieves all multisig transactions where the user is a required signer",
        "tags": [
          "Multisig"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PARTIALLY_SIGNED",
                "READY",
                "SUBMITTED",
                "FAILED",
                "EXPIRED"
              ]
            },
            "description": "Filter by transaction status"
          }
        ],
        "responses": {
          "200": {
            "description": "Transactions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transactions": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/MultisigTransaction"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/multisig/transactions/{transactionId}": {
      "get": {
        "summary": "Get a multisig transaction by ID",
        "description": "Retrieves details of a specific multisig transaction",
        "tags": [
          "Multisig"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The multisig transaction ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transaction": {
                          "$ref": "#/components/schemas/MultisigTransaction"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      }
    },
...
    "/api/multisig/transactions/{transactionId}/submit": {
      "post": {
        "summary": "Manually submit a transaction",
        "description": "Manually submits a transaction that has reached the signature threshold",
        "tags": [
          "Multisig"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The multisig transaction ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transaction": {
                          "$ref": "#/components/schemas/MultisigTransaction"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or transaction not ready"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      }
    },
...
    "/api/admin/transactions/{id}": {
      "patch": {
        "summary": "Update transaction status",
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending_sender",
                      "pending_stellar",
                      "pending_info_update",
                      "pending_receiver",
                      "pending_external",
                      "completed",
                      "error",
                      "refunded"
                    ]
                  },
                  "stellar_transaction_id": {
                    "type": "string"
                  },
                  "external_transaction_id": {
                    "type": "string"
                  },
                  "amount_out": {
                    "type": "string"
                  },
                  "amount_fee": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction status updated successfully"
          }
        }
      }
    },
...
    "/admin/transactions": {
      "get": {
        "summary": "Get all transactions with pagination (Admin only)",
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of transactions"
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}